refactor: moved configs and websocket into helper dir
